Check for EGL_KHR_fence_sync instead of EGL_KHR_reusable_sync
We create an EGL_KHR_fence_sync on the chromium thread
but try to verify if the wait function implementation
is available by testing for the EGL_KHR_reusable_sync
extension in the scene graph thread.
Since the latter extension is not supported by most of
our devices, we never actually waited for the rendering
to finish.

Change-Id: Ied829d4035d42899cfb3f86de018bc48e61c0cc0
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>