UIDelegatesManager: Always print errors if component loading fails.
This setting is useful when developing a new UI set, and one may not initially
know about the suppression of errors, which is a painful problem to run into.

Furthermore, make sure to delete the failed component so that subsequent
attempted use of it doesn't succeed if UI_DELEGATES_DEBUG isn't enabled.

Change-Id: I3d2abc19fb4a007d06795d2e8149ae78d7c9c020
Reviewed-by: Pierre Rossi <pierre.rossi@theqtcompany.com>