Only deploy ffmpegsumo once in debug_and_release
This moves all installs inside the the release-only-with-debug-fallback
block to prevent the debug ffmpegsumo.dll to overwrite the release
one on Windows (and get use for both debug and release builds).

This will use the release version for both instead by default.

This also moves the load(qt_module) call higher to make sure that
build_all is added to CONFIG at that point.

Change-Id: I2584a28e5eaaf5028798703fb387ac2b120d7ea9
Task-number: QTBUG-43348
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>