Fix widgets download state update
Since the DownloadState enum of core does not match the DownloadState
enum of QWebEngineDownloadItem we should never directly compare it,
instead we should convert with the toDownloadState() function,
to fix the missing emission of the finished() signal.
This patch also adds static asserts to verify the match of chromium's
DownloadItem::DownloadState and our core DownloadState.

Change-Id: I51110f55598230e380e15948f75219976d3a68a0
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>