Move navigationHistory out of experimental
Add an offset role to the models.
Implement goBackAndForward that uses this offset.

Also add a complete model, items, that includes current
navigation entry (at offset 0) to allow for the Firefox-style
single menu, and add that one to the nano browser example.

The models are now instantiated lazily as it's unlikely the
three models will be used by the same app.

Change-Id: Ib551738611497c7eb9c501f045cda315968a2ada
Reviewed-by: Andras Becsi <andras.becsi@theqtcompany.com>