Support middle click navigation
When the user does a middle click on a link which originally opens a
new tab (target="_blank"), the newly created RenderView is hidden, so
the new tab has no visible content. We need to unhide it explicitly.

This also adds background tab support for the Quick example browsers
and fixes the profile mismatching warning in the case when we adopt
a new tab without WebContents.

Change-Id: Ia3c2e752a4784d32133cee0ff278c2849b151582
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>