QQuickWebEngineScript: Add a sourceUrl property to complement sourceCode.
QtQuick doesn't have a convenient interface to interact with files, and even if
that were the case, it would either be more awkward (and with increased overhead
-- the price of an extra QObject, property, signal, etc) just to fetch the
contents.

So, we provide a convenience to allow fetching a user script from a file.

Setting the sourceCode directly will reset sourceUrl (if it is set), and setting
sourceUrl will ultimately result in a sourceCode change signal, too.

Change-Id: Iee5abc0d719e2aeeacf1265f695b5a7efee9e0e8
Reviewed-by: Andras Becsi <andras.becsi@theqtcompany.com>