Introduce mock UIDelegates for testing.
Switch JS dialogs test to using this over the testsupport
plugin.

This is a less intrusive approach that lives as self contained qml
modules within the test infrastructure. As an added benefit, it
covers the UI delegation logic from end to end, including what
happens in ui_delegates_manager.cpp.

Change-Id: I1cc0afaf5514d53e1f3b4b0682379dd2d8d8a913
Reviewed-by: Andras Becsi <andras.becsi@theqtcompany.com>