Quick: Share profiles between application windows
The applications should not create new WebEngineProfiles for
each window, because the WebEngineView can not adopt content
from a different profile when the target of the NewViewRequest
is a new window (SHIFT+click) or a dialog.

This also adds a notifier signal to the profile property
in order to avoid "non-NOTIFYable" warnings.

Change-Id: I235789f9bfa1a216f99592204e50266242d0ef1c
Reviewed-by: Andras Becsi <andras.becsi@theqtcompany.com>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>