Fix QtCreator loading of qtwebengine.pro
When QtCreator loads qtwebengine.pro some parsing errors
are shown if it picks up the system python3 for basic detection.
Although we do not plan to support python3 with all of our
scripts (since Chromium does not) these simple cases can
be fixed by using the function version of python's print.
This patch also adds the Chromium source path to the core
INCLUDEPATH so that Chromium headers are indexed as well.

Change-Id: Ia6dec1c776fdcddf875fb8e814d8fb33b7340989
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>