Add cookie API tests
This tests the part of the cookie API that can be tested locally.
The notification about third-party cookies (acceptCookieFromUrl)
can not be tested locally since it requires cookies to be set from
a different domain than the main frame, which requires a remote
host or a local web server with support for virtual DNS.
Testing requires the ability to set cookies for local pages loaded
through the qrc scheme.
We could eventually extend this in the future to enable setting cookies
on registered custom schemes, but for that we might have to implement
our own cookie store.
Since Chromium's cookie store relies on source url's to manage cookies
and qrc:// cookies do not specify a domain, a specific source url is
needed to be able to delete an individual cookie that has been set by
a page loaded through qrc://.
This patch requires a new function on QNetworkCookie to be able to
forward the source url of the cookie to Chromium.

Change-Id: I97dd04b27fbb8ec63060f9b741ad65c29a773a6c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>