Fix quicktestbrowser crash caused by wrong experimental version
The version number of the experimental extension should not
be changed but we should register the latest WebEngineView
for experimental version 1.0. This fixes quicktestbrowser
"Cannot assign to non-existent property "experimental""
issue on startup.
Also update the QtWebEngine import version in the test
browser so that all the new public APIs are visible.
The QML code has already been updated in a previous patch.

Change-Id: Ic8398df6e62ce7cb80c805171f3777d62c0c39a1
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>