Make it possible to run applications with chromium embedded options
This patch adds --enable-embedded-switches command line switch to enable
all the embedded switches on startup instead of during compile time.
Among others this enables pinch-zoom, viewport and impl-side painting,
to make it easy to run an application with these settings even if the
library was not cross-compiled. On embedded these are enabled by default
and can be disabled by --disable-embedded-switches.

Change-Id: I366ed243f1b0e17cbeb900e672d0e15fb4a2b8b0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>