Use v8 backend for proxy resolution
Use the v8 proxy resolution backend instead of the native one in the
multi-process case. This fixes a particular issue on windows where the
winhttp one will attempt to find proxy servers for every single request:
https://crbug.com/40797 . Chrome uses the v8 backend too.

Resort to the native backend though if started with --single-process. This
is because the Chromium renderer and the V8 ProxyResolver can't run in the
same process. See also proxy_resolver_v8.h:

// It is important that *ALL* instances of V8 in the process be using
// v8::Locker. If not there can be race conditions between the non-locked V8
// instances and the locked V8 instances used by ProxyResolverV8 (assuming they
// run on different threads).
//
// This is the case with the V8 instance used by chromium's renderer -- it runs
// on a different thread from ProxyResolver (renderer thread vs PAC thread),
// and does not use locking since it expects to be alone.

Task-number: QTBUG-44763
Change-Id: I7b48cb9d0f3c41fdddffc9eb9f51d83442d80622
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>