Add configurable option for using system FFMPEG
Add qmake command line option to use system FFMPEG and report result
and option in the configure summary. Also checks libraries FFMPEG and
chromium both uses are available on the system since linking will
otherwise fail. Does not check if the version of the libav libraries are
correct, but leaves that to whomever enables the option.

Change-Id: I39c9c47eba19a6e40872a2e9b24a9010cdea1011
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>