Qt Quick Menu: Delay destroying the menu instance
This is necessary on Mac, where we use native menus and, in order
to support catching exceptions emitted from menu handlers, we queue
the 'triggered' signal emission (see qtbase commit 08cc9b9991ae9ab51).
This also means that the 'done' signal is emitted before we get a chance
to receive the menu item's 'triggered' signal, resulting in a noop.

Change-Id: Ie5f06521fb1cdd5def1517a908078c1dd62ab0d0
Reviewed-by: Pierre Rossi <pierre.rossi@theqtcompany.com>