Use the default composition colors for IME composition
We should not set the background color coming from the
input method event since its color scheme might not match
the color scheme of the web page, which can make the
entered text unreadable. Instead we should set the default
as used by the aura and mac implementations of transparent
background and black underline of the current editing context.

Change-Id: I3a4ec5672a917b029fb57cfcc6b36bf93209af2c
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>