Update WebEngineView documentation
The split suggested in the comment does not work: You have to keep the
complete documentation for a QML type in one file, the documentation
bits in qquickwebengineview.cpp were silently ignored so far.

This change moves over the bits from .cpp , adds missing documentation
for new elements and sanitizes the output of others.

Task-number: QTBUG-47909
Change-Id: Ie1ebb5a9fe952c7a76c8fe1b67135614ca90ce7a
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@theqtcompany.com>