fix debug_and_release build on Windows
The public API symbols are compiled in a static library that's linked to core
with --whole-archive when using gcc. There's no equivalent for this in MSVC.
We hacked around this limitation in commit 38944be4 by creating a source file
that includes every public header.
Unfortunately this breaks debug_and_release builds, because we must inject this
generated source file as absolute path, and gyp will generate the same object
file for every source file that's specified with an absolute path...

This reverts commit 38944be4 and replaces the work-around by simulating
-whole-archive on MSVC. This is done by passing all object files that belong
to qtwebenginecoreapi.lib to the linker via a response file that is created
when building qtwebenginecoreapi.lib.

Task-number: QTBUG-48376
Change-Id: Iaa991fe96a2c336d982d29d6924ce950c0be5398
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>