make the render process DPI-aware
On Windows, system calls that retrieve screen properties (like
GetSystemMetrics to get the size of a scroll bar button) are dependent
on the DPI awareness setting of the calling process.
The render process must use the same DPI awareness setting as the
browser process. Retrieve the DPI awareness of the parent process in
the render process and set it accordingly.

Task-number: QTBUG-48380
Change-Id: Ic17d29d0f584e3cf230ac6ea2b08e3aa0d87ccdd
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>