fix python version check
The python code that is used to determine the version number
did not work with python3.
In python3 print is a real function and must be called as such.
Use positional accessors to be compatible with python < 2.6.
Also extend the error message for users that attempt the build
with python3.

Task-number: QTBUG-48507
Change-Id: I49e1fb77c2cc421ac1faed8d8143bf605fbde700
Reviewed-by: Florian Bruhin <qt-project.org@the-compiler.org>
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>