Avoid triggering extra urlChanged signal
Chromium usually clears the pending entry when it fails,
so that an arbitrary URL is not left visible above a committed page.
However, it does preserve the pending entry in some cases,
such as on the initial navigation of an unmodified blank tab.
So in other cases Chromium clears the pending entry
(if the navigation is not in progress), and it does force the UI (URL bar)
to refresh.

In our case we allow failed URLs to stick around in the URL bar
but only when alternate error pages are used.

Change-Id: I26e81e7a99415aa184880dcb647a06d977584119
Task-number: QTBUG-48995
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>