Quick: Close WebUI popups properly
This patch deactivates the web view with a FocusOut event when it turns
to invisible and Chromium closes its WebUI popups.
In case when the window is closing, we can't always rely on the destroyed()
signal or the visibility state of the WebEngineView, since we have multiple
top level windows and the destroying order of the items is not so trivial
as in the single-window case.

Task-number: QTBUG-49099
Change-Id: I802a47c72eed3ed6352f1ba24998622fc95bb48a
Reviewed-by: Alexandru Croitor <alexandru.croitor@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>