Add support for checking if audio is played in a page.
Add support for checking if audio is played in a page, as well
as the ability to (un)mute the audio.

Modify demobrowser example to show (muted) in the tab title, if
the tab is muted, or (audible) if there is audio playing in the tab.
Fix HTML5 audio/video (un)mute to also work.

Change-Id: I7213645e67be2f9da1c5f96cdf6c7eef5341ae4b
Task-number: QTBUG-48788
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>