Make lazy contentsAdapter initialization lazier
By postponing the creation of the contents adapter until the eventloop
is started we can avoid creating it multiple times if the profile is
changed during QML loading.

This also works around a bug on startup of quicknanobrowser in single
process mode, but doesn't solve the underlying problem.

Change-Id: I8859fd91dd828658b363ebf89ca619a193deaefc
Task-number: QTBUG-50256
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>