Fix doubled characters when using IME on Windows
On Windows, when exiting the IME by pressing a cursor key or clicking
into the edit field, we do not receive an QInputMethodEvent with a
commitText. Instead, we get an empty QInputMethodEvent and a
key{Press|Release}Event pair with a key of zero. The committed text
is in the event's text property.
Do not call ImeConfirmComposition for the empty QInputMethodEvent
but for the later keyReleaseEvent. Do not forward those key events.

Change-Id: I844aa05493aca4c388a8b1de835baf2a819558f5
Task-number: QTBUG-50252
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>