Add API for context menu data
For QWebEnginePage users to be able to make custom context menu or
extend the default context menu, they need information about the
context.

This patch adds a QWebEngineContextMenuData class that contains
contextual information when a context menu is being requested.

This is also means we can finally generate a custom context-menu in
the demobrowser the same way we did in QtWebKit.

Task-number: QTBUG-51079
Change-Id: I695b8cbf648bdc7f119b0ed51ab685cf2f8de8e4
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>