Initialize lazy base::CPU when IO are allowed
We should initialize base::CPU to get it a chance to read /proc/cpuinfo
when IO operations are allowed. So it prevents the following error
on ARM platform:
```
[0204/071425:FATAL:thread_restrictions.cc(38)] Function marked as
IO-only was called from a thread that disallows IO!  If this thread
really should be allowed to make IO calls, adjust the call to
base::ThreadRestrictions::SetIOAllowed() in this thread's startup.
```
In the same manner it is done in ChromeMainDelegate::PreSandboxStartup.

Change-Id: I1facd9d856a8dbfdf03ddbb1ab8d76d2efe69050
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>