Fix CXX :visited selector
This patch fixes highlighting of visited links and now Acid3 test
completely works (100 of 100).
The reason is that VisitedLinkMaster must be initialized before a new
RenderView will be created. Otherwise, it will not handle
content::NOTIFICATION_RENDERER_PROCESS_CREATED event and therefore
VisitedLinkSlave will be left uninitialized (salt_ is zero). Because of
this reason CSS :visited selector was broken and didn't work.

Change-Id: I769cd5dbae2ffb95fd128df634a54e562b9cc91d
Reviewed-by: Ilia Kirianovskii <ilia.kirianovskii@lgepartner.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>