Fix crash when link opens a modal QDialog, using the trackpad.
When a QWebEngine link is clicked on, and as a result a modal dialog is
opened, a QEvent::TouchCancel without any touch points is forwarded
to Chromium, which tries to access the first touch point
position, and causes a QList assertion.

Fix consists of two parts:
1) Make sure that no TouchCancel is forwarded, in case if no TouchBegin
or TouchUpdate was issued beforehand.
2) Because QEvent::TouchCancel events might contain an empty touch
point list, and Chromium expects at least one point, make sure
to forward the last saved touch points (saved in previous TouchUpdate)
together with the TouchCancel.

Task-number: QTBUG-48661
Change-Id: I1eeb2980417b1b04e8387dc9f82f935ef2bd8f00
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>