Fix CDM plugins not always being loaded
This fixes the issue in a similar fashion to what Chromium is doing
in chrome/browser/download/download_target_determiner.cc.

PluginService::GetPlugins is the only way that the plugin info can
be loaded. Since we can't do it on demand from the IO thread in a
synchronous handler, do explicitly during the WebEngineContext
initialization.

Task-number: QTBUG-50132
Change-Id: I8ed796b541c880caf8546230b3a5011047eb0fe3
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>