Add support for compiling WebEngine with Clang sanitizer options
This change pipes the already supported Qt sanitize options to
Chromium, so the Chromium build system adds the correct compiler
sanitizer flags.

The change also checks whether the requirements to build a sanitizer
enabled Chromium are met (currently works properly on macOS with an
Apple clang >= 7.3.1, and on Linux with clang + libc++).

Change-Id: Ibab2d5735935ddc92226575fc0e9f5912b807329
Task-number: QTBUG-52656
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>