Get additional sanitizer dependencies from upstream chromium.
The buildtools directory contains custom libc++ libraries, which are a
second-level git submodule (the first-level being the buildtools
submodule, and the supermodule being chromium itself). The
init_repository.py script does not support checking out submodules
recursively, thus leading to the mentioned libraries being skipped.

This change adds a hardcoded addition to make sure the buildtools
submodules are checked out when init_repository is run, as well as
support for copying all nested files in the submodule tree, when the
take_snapshot script is executed. Thus it copies all necessary
sanitizer dependencies.

Change-Id: Icb5cf5b52c3d0a83c4690fb7d641cee4f5dc1132
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>