Demobrowser: Fix black screen after exiting fullscreen mode on OS X.
When exiting fullscreen while viewing a full screen video, the video is
replaced with a black screen, and does not switch back to the initial
page.

This happens when hide() is called on a native window while being in
fullscreen, which hides the window, but does not exit full screen mode,
thus showing a black screen.

The workaround fix consists of two parts:
1) Make sure to delete the fullscreen window object, to release the
OS virtual screen it occupies (visible in mission control as a black
rectangle).
2) Hide and show the main page window, to force exit fullscreen mode.

Change-Id: I364d4427e84957ed97f57a14e2cfc2c95aed6f48
Task-number: QTBUG-52058
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>