Adds qwebengine_convert_dict tool
This is a tool which converts hunspell 'dic' dictionaries
to binary 'bdic' format, which is required by spellchecker
in chromium. Tool is compiled as a qt console application.
It reuses linking information produced by gyp.

This tool cannot be used in cross-builds due to limitations
of the gyp-based build system.

Change-Id: Ibee8cbc6048b522c0e4fe22b21c91c649c8515bc
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>