Reset the selected text of a page when setHtml is called.
When text is selected inside a QWebEngineView, the selection is stored
in a RenderWidgetHostViewBase instance. When QWebEngineView::setHtml is
called, the stored selection is not cleared, and thus requesting for
the selected text will return stale data that is not present in the
page anymore.

Fix consists in calling WebContentsImpl::Unselect() after the new
html is loaded and focused in WebContentsAdapter::setContent().

Change-Id: Idd0f3187f324863b9a805af6a288dccfcbd5566f
Task-number: QTBUG-53033
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>