Fix QtWebEngineCore::initialize() for Qt < 5.6.1
In Qt versions < 5.6.1 the constructor of QCoreApplication directly
calls QCoreApplication::init and QtWebEngineCore::initialize() in turn.
That means that qobject_cast<QGuiApplication*> cannot work, because the
object is not fully constructed yet.
This was fixed for Qt 5.6.1 in commit qtbase/1b441c39.

This patch turns off the fix for QTBUG-51789 when building QtWebEngine
against Qt versions older than 5.6.1.

Task-number: QTBUG-51789
Change-Id: I092fb559ea4449bd443737d5962b4b87dee88cdd
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>