Update scroll information in touch events
Touch events have causesScrollingIfUncanceled flag that is used by
TouchEventQueue::TouchMoveSlopSuppressor for determining whether they
can be suppressed or not. Previously it was always false what caused
the events be suppressed. Then, it was set to true what caused the
events cannot be suppressed event if they can. Now the flag can be
dinamically changed depending on the result of
ui::FilteredGestureProvider::OnTouchEvent.

Change-Id: I625a5c5d9b7597e25406e55c461efb7187e7c482
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>