Disabled WebEngine views should not receive focus.
Currently if a QWebEngineView or a QQuickWebEngineView is disabled
using setEnabled(false), after loading a web page, the views are
automatically focused, and a user might see a blinking caret in an html
input for example, even though the user can't interact with it.

Fix consists in not calling the Focus() method whenever a view is
disabled.

Change-Id: I1014fb5898a5ddf01a4e9b14c3eaf5d4006e5131
Task-number: QTBUG-53159
Reviewed-by: Michal Klocek <michal.klocek@theqtcompany.com>