Fix event forwarding when activeFocusOnPress is false.
If a QQuickWebEngineView does not have focus, and activeFocusOnPress is
set to false, a user can still partially interact with the view. For
instance hovering the mouse over a link would change the cursor, a link
can be clicked to go to a different page.

Clicking on a text input field would focus the text field, but entering
characters will not be possible, because the view does not have QtQuick
keyboard focus, and clicking does not give the focus (because
activeFocusOnPress is set to false) and this leads to confusing
behavior.

Thus the fix is to make sure no mouse / keyboard events are forwarded
to Chromium if the view has no focus, and activeFocusOnPress is set
to false, in order to maintain a more user-friendly behavior.

Manually forcing the focus via some user-provided method that calls
forceActiveFocus() would allow further proper interaction.

Change-Id: I72c3ff69438972b9a93ee2d415fa1d4b44b86cd9
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>