[Simplebrowser] Fix command line option parsing.
Passing just command line options to Chromium prevents loading of the
initial startup url. This happens because the startup url is
overridden by the last command line argument, which happens to be a
Chromium option, thus an invalid URL.

Fix consists in using the last command line argument as the loading url
only if it can be parsed as valid url.

Change-Id: I3615ef57b6eddc5098de79cba1b1ba6b6af193ef
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>