Adaptations to Chromium 51
A few remaining details need to be changed to port to 51:

- AX_STATE_FOCUSED no longer exist and is now queried from the
accessibility manager.

- OnDuplicateSection no longer exist and is no longer needed on
Windows.

- Vorbis and AAC codecs are no longer included as supported
in chrome_content_client.cc so we match that.

- Blink is no longer started lazy and the method WebKitInitialized no
longer exists in RenderProcessObserver. We must therefore get rid of
the observer, and can mark qrc immediately.

- Chromium will now use gpu memory buffers for video by default on both
Linux and OS X. It however no longer works for us on either platforms,
and since it doesn't give us any performance benefits, we might as well
disable it.

Change-Id: Ie0ee361f05a1e2207fcc5b88f98f9026ccd45f78
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>